Move Provider to its own module
